fix allocation sizes for `StringBuilder`
While mainly a fix for previous commits, this adds a few safety checks
and optimizations.

In cases of inserting even 0 bytes an allocation would be created with a
single null byte. This makes it so an allocation is only made if bytes
are actually being inserted.

NULL checks have been added for realloc as it may return a NULL pointer
if it is unable to allocate. Instead of UB it will now panic.